MB-37068 New timer lag process
In preparation for removing timer2.erl this process keeps track of timer
lags by sending itself a message every second.  And on receipt of
the message it records the difference betwen the time of receiption and
the expected time.

Currently this uses the same system_stats_collector but will move to use
something else in a subsequent change.

Change-Id: I1b7d353dd54eb7651d88b18e71994f192c0726f7
Reviewed-on: http://review.couchbase.org/119875
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>