MB-37514: Fix dump-guts to return correct error code in case of error
It used to call init:stop/1 before, but the problem with init:stop/1
is that it is asynchronous. After calling init:stop the script
continues and finishes earlier than init:stop is able to interrupt
it. That's the reason why script always returns 0, even if error
occurs.

Change-Id: I286f35c9463517478656304215844993d6994f02
Reviewed-on: http://review.couchbase.org/120504
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>