MB-37447 Detection of skipped messages
When send_iterval was used the handle_info function would be able to
use misc:flush to determine if and how many messages were being
flushed/skipped.

When send_interval was changed to use send_after we lost the detection
of the skipped messages.  This change implements the ability to detect
and report the skipped messages using erlang:send_after with an absolute
time.  On receipt of the message we calculate the the difference between
the time of receipt and the expected time.  This difference, if larger
than the time interval, is reported as skipped messages.

Change-Id: I80f6b889d89fd4d60f8d61e3df2f4c944fb7d131
Reviewed-on: http://review.couchbase.org/c/ns_server/+/120673
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>