MB-37269 Change memcached admin_user on upgrade
When the change was made to use "@ns_server" instead of "_admin" as the
memcached admin_user the upgrade code was missed.  As a result nodes
which upgrade from prior to 5.0 have "_admin" and nodes which never ran
pre-5.0 use "@ns_server".  Because of this filtering of audit events
could only be done for one or the other depending on which node the REST
API "settings/audit -d disabledUsers..." ran on.

Change-Id: I2386d81c4d90da7fd574fd89047af37aa552b969
Reviewed-on: http://review.couchbase.org/121225
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>