MB-32401 Optionally disable "WWW-Authenticate" response
By default, the ns_server REST API responds with a WWW-Authenticate
header to unauthorized requests which cause the browser to prompt the
user for username/password.  These credentials may be cached by the
browser and may make the cluster vulnerable to a CSRF attack

This change provides a way to disable the WWW-Authenticate response.
Disabling is done using couchbase-cli:

./couchbase-cli setting-security -c localhost:9000 -u Administrator \
        -p asdasd --disable-www-authenticate 1 --set

which is implemented in http://review.couchbase.org/#/c/121821/

Change-Id: Ib10c524a09a27a34f229e507f5f370345a9aa943
Reviewed-on: http://review.couchbase.org/121288
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>