MB-36577 - N1QL formatter fails on certain queries
This bug was caused by two related problems. The first issue was that
indentation was increased or decreased if the level of parens changed,
but the magnitude of paren depth wasn't considered. E.g., it increased
indentation regardless of whether the paren depth increased by one or
ten. The related bug was that the stack of indentation could get
popped while empty, leading to an intentation string of "undefined".

The fix changes the indentation proportionally to the change in paren
depth, and adds a sanity check to prevent popping the indentation
stack when empty.

Change-Id: I4313456d6f8bb6575e5b31209cff1d427f980aa1
Reviewed-on: http://review.couchbase.org/121681
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>