MB-37801: Prevent over-ride of rebalance report files
During log redaction we create new redacted files in a temporary
directory. However, if the name of the file to redact is given as an
absolute path, the output file and input file are the same, this is
because in os.path.join, if a component is an absolute path, all
previous components are thrown away and joining continues from the
absolute path component.

This results in over-riding the server files.

The solution is to use the basename of input file to generate the output
filename.

Change-Id: I5b57f723250c5827dab4bb2c456a638c3587a952
Reviewed-on: http://review.couchbase.org/121689
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>