MB-38048 couchbase-server check for Couchbase running
This change adds the --status, -s option to couchbase-server which will
print whether or not couchbase is running.  This is determined by
getting the pid from the couchbase-server.pid file and then checking to
see if that process is running and is named "beam.smp".

The script returns the following exit status codes:

0       program is running
1       "pid" file exists and program is not running or not beam file
3       program is not running

This patterns after

 http://refspecs.linuxbase.org/LSB_3.0.0/LSB-PDA/LSB-PDA/iniscrptact.html

Change-Id: Ic080d376bad33a516e0de5a905115dd4ecdbb352
Reviewed-on: http://review.couchbase.org/122743
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>