MB-37802: Change LDAP server cert text area label
It now reads "LDAP Server Root Certificate" instead of "Certificate Text".
Users have been confused as to whether the certificate to be pasted in the
text area is the LDAP server root cert or some kind of client cert to be
used by ns_server when connecting to LDAP. This is hopefully a little
clearer.

Change-Id: Ia50762747273f4ca537cd53ac5776239d4738d89
Reviewed-on: http://review.couchbase.org/123740
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>