MB-38322 restructure memcached_permissions code to be able to
...create multi level json for memcached permissions, placing them
on bucket, scope or collection level

the mechanism is applied on bucket only level in this commit

The resulting json is different from the pre-commit version in the
following fashion:

1. List of permissions now appears under privileges subkey of the
bucket, not at the bucket key level

ex:
"user": {
    "buckets": {
      "test1": {
        "privileges": [
          "SimpleStats"
        ]
      }
    },
    "privileges": [
      "Stats"
    ],
    "domain": "local"
  }

2. buckets without any permissions are omitted

Also the caching of calculated permissions for individual roles
is dropped. I plan to work on performance optimization later.

Change-Id: I40f9dd9ac8e58ac21e4ee795866df546dd65fd44
Reviewed-on: http://review.couchbase.org/c/ns_server/+/124925
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>