MB-38487 Handle truncated local failover log.
Failover log is capped at 25 entries in kv-engine. Once the limit is
reached, old entries in the log start falling off.

When determining if two vbuckets are compatible during delta recovery,
the code handled this situation when dealing with the failover log
from the new active vbucket, but didn't do the same when dealing with
local failover log.

Alas, if the cluster doesn't receive any mutations and undergoes many
failovers or dirty restarts, eventually enough failover entries will
be created and all the mutations will be from "prehistoric" times,
that is, there'll be no way to identify the failover UID for those
mutations.

In the referred ticket, the failover log for one of the affected
vbuckets looks as follows:

  [{224004494071701,1},
   {192067325548961,1},
   {245575217719652,1},
   {51175631673048,1},
   {49035851937624,1},
   {271774423209112,1},
   {101210916020687,1},
   {112566719948469,1},
   {12269405125050,1},
   {64464211619117,1},
   {218513513177021,1},
   {50078031322354,1},
   {64925328152668,1},
   {274501948063656,1},
   {26679072113212,1},
   {82742009590079,1},
   {93313425930461,1},
   {204212430706959,1},
   {197749909789638,1},
   {17138748499922,1},
   {189076684079652,1},
   {216799193295675,1},
   {171173845852830,1},
   {175869414126129,1},
   {76081282101286,1}]

While the high seqno is 1. So the mutation at seqno 1 happened before
any of these entries were created.

Change-Id: I6385653cf269b16f66ac7195f5b7dbecae5efb02
Reviewed-on: http://review.couchbase.org/c/ns_server/+/124546
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Wayne Siu <wayne@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>