Add warning comment to avoid future upgrade breakage
The get_current_version/0 function provides the version of the config
and is used during offline upgrade to determine what upgrade steps need
to occur.  If there are shipped releases that don't recognize the
config version an offline upgrade will fail.  This change adds this
information as a comment in the code.

Change-Id: Iee240ebb5b400888e6b4c8765c039a285c87607b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/124548
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>