MB-38698 - improve language for 'update notifications'
Because of file location changes, need to re-check-in to master.

The 'update notifications' checkbox and tooltip needed improved
language to explain what the checkbox actually does, and what data is
collected as part of the update notification.

Also, move the setting to earlier in the wizard to make it more
obvious to users.

Change-Id: Ia4cfa6faee7d05994790cacf1dce3f4c3202468b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/125597
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>