Construct pluggable-uis.js from modules in pluggable configs
In 6.5 and earlier pluggable UI components were able to register with
the Admin UI app when they were loaded. With the move to modules, this
is no longer possible as modules are loaded asynchronously and we now
need to explicitly load pluggable modules from the main app.

The way we accomplish this is by adding a new 'module' attribute to the
pluggable configuration. This information is used to generate a
"pluggable import" module which we call pluggable-uis.js. This import
module looks as follows:

  import pluggableUI_cbas from "/_p/ui/cbas/./cbas.js"
  export {pluggableUI_cbas}
  import pluggableUI_n1ql from "/_p/ui/query/ui-current/query.js"
  export {pluggableUI_n1ql}
  import pluggableUI_fts from "/_p/ui/fts/./fts.js"
  export {pluggableUI_fts}
  ...

This allows the the main app to import the pluggable modules as follows:

  import * as pluggableUIs from "/ui/pluggable-uis.js";

  angular.module('mnAdmin', [
    ...Object.values(pluggableUIs),

As I was doing this work, it became apparent that the pluggable configs
have gotten stale in some respects. E.g. most components still use the
old form of 'rest-api-prefix', some components still refer to 4.5 UI
compat version. These issues we should address in subsequent checkins
in the pluggable components.

Change-Id: I537f0189de27547cc9020ebeda4690042f2a3bba
Reviewed-on: http://review.couchbase.org/c/ns_server/+/125551
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>