[bp] MB-38715 Report /proc/vmstat's allocstall
On linux versions up until 4.10 this is /proc/vmstat's allocstall stat.

On 4.10 onwards it's the sum of /proc/vmstat's allocstall_dma,
allocstall_dma32, allocstall_normal, allocstall_movable.

For non-linux systems it'll return -1

This change requires the following manifest change in order to bring in
the corresponding sigar change:

http://review.couchbase.org/c/manifest/+/125748

Original change:

http://review.couchbase.org/c/ns_server/+/124033

Change-Id: Ibde53205101b41427733318387a476e33917c14e
Reviewed-on: http://review.couchbase.org/c/ns_server/+/125749
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>