MB-38443 Ignore "unimportant" items in version hash
In order for the UI to display index status a "last scan time" was added
to the response to /getIndexStatus which ns_server uses to get status from
the indexer (MB-35197). This information is obtained by the ns_server
service_status_keeper infrastructure every five seconds. If the information
has changed (via a hash of the info) then notifications are sent to clients
watching for indexChanges. The "last scan time" changes for every n1ql query
that uses a primary or secondary index (MB-36086). We suspect this is the
reason the streaming clients are un-hibernating every five seconds.

This change doesn't include the "lastChangeTime" or "progress" in the hash
of the info.

Change-Id: I5d6cb8c18103e15a4530a96ad88d5b08f57fd152
Reviewed-on: http://review.couchbase.org/c/ns_server/+/125751
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>