Allow pluggable components to use default port for service
The old rest-api-prefix attribute of the pluggable UI config
is deprecated and no longer used. However, it was convenient in
that it allowed pluggable UIs to not be required to specify the
service port - something that the new rest-api-prefixes
attribute does not allow. This patch changes the way the
rest-api-prefixes are handled allowing pluggable UIs to not
explicity specify a portName, which means the defautl port for
the service will be used.

Change-Id: I9e69bd6932bed7d4aba35185e5111c7950a48c05
Reviewed-on: http://review.couchbase.org/c/ns_server/+/127903
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>