MB-33812 Add session timeout audit event
ns_server has a session timeout but the UI periodically "refreshes" so
the timeout isn't hit. The UI can be configured to auto logout after a
specified number of minutes.

But in the case where the user closes the tab the session will timeout
after ten minutes but no audit is generated.  This set of patches will
resolve this.

This patch defines the audit event.

Change-Id: I29717ef3b170f8893e2335d0dfa39161a5755b6d
Reviewed-on: http://review.couchbase.org/c/ns_server/+/126871
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>