Add "is module prefix" to pluggable UI config
Prior to this patch, pluggable components couldn't have more than
one REST API prefix and have the pluggable UI work as there was
no way for ns_server to determine which of the REST prefixes
associated with the plugin allows retrieval of the pluggable UI
top-level JavaScript module.

Currently, no pluggable component has more than one REST API
prefix - however it used to be the case that Analytics had two
prefixes so it is possible and it may be needed again.

With this patch pluggable UI components can identify the REST
API prefix to use to serve the module as follows:

    "cbas": {
      "portName": "cbas_http_port",
      "isModulePrefix" : true
    },
    "cbas-admin": {
      "portName": "cbas_admin_port"
    }

isModulePrefix can only have the value true (or be undefined)
and it's not permitted to specify more than one prefix to be
a module prefix. There's no need to specify which is the  module
prefix if the plugin only has one prefix.

Change-Id: I201ada23d6f5cc87626193b5716b819d6433df12
Reviewed-on: http://review.couchbase.org/c/ns_server/+/127905
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>