MB-39028 Disable newline translation.
When reading and writing files for the purposes of redaction.

This addresses the immediate issue of binary files being corrupted by
redaction and it should be bug compatible with the behavior prior to
the transition to python3.

This is only meant as a short-term solution for 6.6.

Change-Id: Ibb242cd8858477e86dc669a54826bbf5a1481f45
Reviewed-on: http://review.couchbase.org/c/ns_server/+/127942
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>