MB-38552 Enforce maximum number of scopes and collections
Define the maximum number of scopes and collections but allow for
changing via parameters in the config.  The maximums are cluster-wide
total limits for all buckets.

When doing verification of the operation check that the maximum number
of scopes and collections doesn't get exceeded.

Change-Id: Ibd0d78976cc9681f0cd72d1e2c2611185616e40b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/128224
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>