Don't return last minute stats when node starts ...
... for hour, day, week, month and year stats

This is essentially a hack that is needed for backward comaptibility
on mixed clusters. It helps to work around the issue that leads to
showing a lot of gaps on last minute graphs when new cluster node
starts/joins the cluster. The issue is caused by the fact that
pre-7.0 code assumes that if node has "hour" stat for the last
minute, it should also always have "minute" stats for the last
minute, which is not true for backward compatibility code in 7.0 node

Change-Id: I74d0fd55cfcd6e8594945a9d51ad38e96f0b6a42
Reviewed-on: http://review.couchbase.org/c/ns_server/+/135472
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>