Change UI compat version to 7.0
We've made a lot of changes in the pluggable UI framework while
concurrenlty changing the code in the pluggable UI components
themselves, so upgrading the UI compat version isn't strictly
necessary. However, it seems a reasonable practice to upgrade
the compat version to reflect the fact the significant non-
backward compatible changes have been made.

Change-Id: I59095279b039d6882105b4c96fca0dfb38b36302
Reviewed-on: http://review.couchbase.org/c/ns_server/+/129572
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>