MB-39648 Correctly log DCP GetErrorMap command
The following DCP command will be logged with its name instead of
"not_dcp".

    GetErrorMap = 0xfe

Some notes from the KV team:

When a consumer/producer connect, the consumer is sending a GetErrorMap
command to the producer to try and determine if the producer is 5.0 or
greater. The way KV sets up the GetErrorMap request means that enoent is
possibly expected... For the dcp_proxy -  we should still log this .
It is only one message each time the producer/consumer connect. The response
is the basic memcached response of 24 bytes - same as stream_end.

Change-Id: I78520620af071c73f9f3b86bfa12a94db4b3e446
Reviewed-on: http://review.couchbase.org/c/ns_server/+/129850
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>