MB-38339 Basic auth for prometheus
When ns_server starts prometheus we generate a random password and then
hash it using pbkdf2.  The resultant hash as well as the salt and
number of iterations used to generate the hash are written to a file
whose name is  passed as an argument to prometheus.

A user of a prometheus REST api must pass the random password that was
used.  Prometheus authentication will see if that password along with
the salt and number of iterations results in the expected hash.

The corresponding changes to prometheus are merged into:

https://github.com/couchbasedeps/prometheus/tree/couchbase-v2.17.2

Change-Id: Ia418be167a66e1d1f412e95fa9a7e171edc7edd3
Reviewed-on: http://review.couchbase.org/c/ns_server/+/130090
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>