MB-40641 time offset "in sync state" could repeatedly toggle
The time offset detection code monitors the time difference between
a non-master node and the master node.  If that difference is above
a fixed threshold, we consider the node's time to be out of sync.
If the time difference is very close to that threshold, a node could
cycle between "in sync" and "out of sync" state due to jitter in
observed time deltas.

This problem is fixed by introducing hysteresis: the "out of sync"
state is made a little stickier than the "in sync" state.  This is
implemented by ensuring that the time difference above which a
node is considered to be out of sync is greater than the time
difference below which a node is considered to be in sync.

Change-Id: Ib48d6bcfac23124ec615a690c30358077283e0be
Reviewed-on: http://review.couchbase.org/c/ns_server/+/133488
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Sam Cramer <sam.cramer@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>