MB-40728 Specify server name in proxy check
When doing a log colletion from the UI and a proxy is specified we
attempt to validate the reachability of the proxy.  Prior to this change
the "Client Hello" done as part of the TLS handshake protocol was not
specifying the server name indication (SNI).  The SNI allows multiple
hostnames to be served over HTTPS from the same IP address.  Without the
SNI the server cannot determine which hostname is being specified and
thus the handshake fails.

This change uses the proxy_ssl_option to specify the SNI as an option to
lhttpc:request/6.

Change-Id: I26e160314119798b0b13ea3ed7fa99f56320f976
Reviewed-on: http://review.couchbase.org/c/ns_server/+/134270
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>