PromQL formatter: Merge 'or' operands when possible
Make formatter to do the following transformation automatically:

"f({name=`m1`, ...}) or f({name=`m2`, ...} or ..."
change to
"f({name=~`m1|m1|...`, ...})"

The main reason for doing that is that it works faster this way
while it is often easier to generate separate subqueries first
and connect them via 'or' later.

Note that it's correct only if function 'f' commutes with 'or'.

Change-Id: Ib122e6c22afc180c207a67472c9833dcbdc6a6e4
Reviewed-on: http://review.couchbase.org/c/ns_server/+/135201
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>