MB-40966 Don't access private ets tables
diag_handler should not attempt to access a private ets table
as it leads to a badarg error.  Instead let the user know the table is
inaccessible.

Specificallly, http://www1.erlang.org/doc/man/ets.html says

In general, the functions below will exit with reason badarg if
any argument is of the wrong format, if the table identifier is
invalid or if the operation is denied due to table access rights
(protected or private).

Change-Id: I1cc14f4a9017746b8beedf9210aa9c29144ec098
Reviewed-on: http://review.couchbase.org/c/ns_server/+/135585
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>