MB-41068: support the redistributeIndexes setting
This is a new indexer setting being introduced in Cheshire Cat and
it shouldn't be settable in mixed mode clusters where the cluster
compat mode is < 7.0. During upgrade this gets added to the indexer
setting config and defaults to false.

Note that the indexing team is considering changing the default to
true, but haven't fully decided at this point.

Change-Id: I41e184d4f996a190fb3881bb1a282970d8b2f6fd
Reviewed-on: http://review.couchbase.org/c/ns_server/+/136164
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>