MB-41439 Provide cbimport with a cbauth environment variable
Providing cbimport with a cbauth environment variable allows cbimport to
authenticate via cbauth instead of using the bucket password. This is a
more proper way of authenticating and allows our REST client to use the
correct credentials when authenticating to services on other nodes in
the cluster.

Change-Id: I48f6aba15ffc6e8a7d41c3fd3f5e20ff06b20e2f
Reviewed-on: http://review.couchbase.org/c/ns_server/+/136704
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: James Lee <james.lee@couchbase.com>