MB-41812 Preserve directories across node removal
When a node is removed from the cluster we will now retain the cbas
directories and eventing directory.  This allows a user to specify a
custom directory and have it persist across a node removal/node add.
This is consistent with the current treatment of the data and index
directories.

Change-Id: I2aa2508b33804d721f47b6c8f9783bc1fdcce1a9
Reviewed-on: http://review.couchbase.org/c/ns_server/+/137445
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>