CBD-3697: PRODUCT_VERSION is not generally cached
PRODUCT_VERSION is now only a cache variable if the developer explicitly
specifies it at the command line (see tlm/CMakeLists.txt). If it is not
so specified, tlm will set it non-cached based on the version in the
manifest, or "0.0.0" if this cannot be determined. In any case, it will
always be set, and ns_server should not change the decision of whether
to cache it.

Change-Id: I2e189e80e95fc834d1208f8a077ed04d99132ff8
Reviewed-on: http://review.couchbase.org/c/ns_server/+/137801
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>