MB-41765: User should be able to view/manage collections
according to their permissions.

The permissions can be obtained per scope through scope_admin and for
the entire bucket through bucket_admin.

scope_admin can drop the scope it manages but not create it, similar to
bucket_admin which can drop the bucket it manages but not create it.

Change-Id: I0548d415be413f2a63e36ac7c9f805d1bbb873a6
Reviewed-on: http://review.couchbase.org/c/ns_server/+/138326
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>