MB-41022: Support disabling UI alert pop-ups (infrastructure only)
Adds a list of alerts ("pop_up_alerts") which trigger UI pop-ups to
the "email_alerts" configuration stanza.  Adds upgrade handling for
that list.

Extends the "/settings/alerts" REST API so it can be used to get and
set the list of alerts which cause UI pop-ups.

Does not include changes to UI for disabling alert pop-ups.

Change-Id: Ia3b441d986d55351b9a217a0030a7798401b938f
Reviewed-on: http://review.couchbase.org/c/ns_server/+/134559
Tested-by: Sam Cramer <sam.cramer@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>