MB-42218: should have one time_out_of_sync key after upgrade
When a cluster is formed, we do an online upgrade, which upgrades
the configuration starting with the lowest possible cluster version,
regardless of a node's version.  This happens when we are upgrading
a node running down-rev software and when we are forming a cluster
composed of nodes running up-rev code.  In the latter case, a node's
configuration may contain information which is up-rev relative to
the cluster version being upgraded.  Consequently, the functions
used to perform the online upgrade must ensure that they are not
adding configuration information which is already present in the
node's configuration.  If this is not done, we can end up with
duplicate information in the configuration.

Change-Id: I5e2ee929ba4601e982390c06adf724a14def753b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/138737
Tested-by: Sam Cramer <sam.cramer@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>