MB-42689 Use new heartbeat_interval when changed
When the mb_master heartbeat_interval changes we'll now use the changed
value. Prior to this change we used timer2:send_interval which sends the
message at the specified period. But use of that function is deprecated
in more current releases and so we're switching to erlang:send_after to
match the current releases.

Change-Id: Ie2d11d892997b4610be8ade8371266c2b1f9b286
Reviewed-on: http://review.couchbase.org/c/ns_server/+/140135
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>