MB-43105: Suppress logging of system-event messages
Noted that when using the bulk collection API the replica
was lagging far behind the active after a 'large' manifest
was deployed. This was traced to the logging of DCP system
event messages.

The following are very rough measurements but a 3 node
cluster_run took ~5 minutes to get 1,000 collections to all
replicas (terminal logging). 6 seconds with this suppression.

A 2 node 'rpm' build (i.e. logging to file) took ~3 minutes.


Change-Id: Ifd602723f2c4d82309d344a7da87834a6932e55a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/141645
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>