MB-39609 - Can't drop index from UI
With the advent of collections, the syntax for dropping indexes has
changed from:

DROP INDEX <bucket name>.<index name>;

to:

DROP INDEX default:<bucket name>.<scope name>.<collection name>.<index name>;

There is a second syntax for dropping indexes, but only the above
syntax is compatible with both 6.5 and 7.0, so it's safe to use in
compat mode.

A second part of the bug was that the errors resulting from the
invalid syntax were not being surfaced to the user. The change to
mn_gsi_item_details.js makes any errors visible.

Change-Id: I1accb6350977be46b139ef8ea94b835fafdbfe5d
Reviewed-on: http://review.couchbase.org/c/ns_server/+/141727
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>