MB-43196 During config reload try connecting to kv on all afamily
and use the connection which works.

When we attempt to reload memcached config, we may have a situation
where the address family in ns_server has changed but not propogated to
memcached yet. We might fail to communicate with memcached over new address
family as memcached may not be listening on the new address family. This
bug is especially seen when we go from ipv4-only to ipv6-only and vice
versa.
Since we cannot know for sure which address family memcached is
listening on, due to various error paths that can lead to mismatch in
what ns_server perceive the memcached config is and what in fact is
applied in memcached, it is simplest to attempt connection using both
address family.

Change-Id: Id8fb97470bbd2bdb4573c8a01b16650cc652601b
Reviewed-on: http://review.couchbase.org/c/ns_server/+/142182
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>