MB-42373, MB-41189: Add endpoint to disable unused cb_dist ...
... listeners

In some situations, when changing net settings we don't know
the previous state of the system. It makes disabling of unused
interfaces problematic. The new api just closes all the listeners
that are in use. It will be used by UI for correct node
initialization and by CLI for changing ip-family and n2n encryption.

Change-Id: Ic6a66d72df4a15a923497035eab462191a83b82a
Reviewed-on: http://review.couchbase.org/c/ns_server/+/142449
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>