MB-43207 DeDup projector port in returned REST response
The projector uses the same port number for SSL and non-SSL. As a result
there's no way to say with certainty...return the SSL port or return the
non-SSL port and guarantee the right value. The bucket info thus would
return "projector" with the same value twice.  This change deDups it.

Change-Id: If1a8e7b61a24d4b5bb7896f60d44b64047a42285
Reviewed-on: http://review.couchbase.org/c/ns_server/+/143063
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>