MB-43720: Handle IOErrors from zipping files
At the start of cbcollect we build a list of ("collect") files we zip.
Such a file may not exist when we later try to zip it.  This has been
seen with old rebalance reports, which are purged on a regular basis. In
theory it could occur with any file we are collecting.

When we encounter this scenario we should emit an informative error
message and drive on, continuing to zip the remaining collected files.

Change-Id: I1342a8670cc6988fc38f99dd52b5ec03060090bc
Reviewed-on: http://review.couchbase.org/c/ns_server/+/143759
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>