MB-43849 dump-guts assumes all services available.
dump-guts used to look at ns_config for the list of services on the node
but the information is now written to chronicle. So as a workaround, we
assume all services are on all nodes. This will be undone when dump-guts
is able to read from chronicle.

Change-Id: Ied8b254937c830bc29cdc92189ebc48e94bf208d
Reviewed-on: http://review.couchbase.org/c/ns_server/+/144314
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>