MB-43269 Remove eventing_settings_manager upgrade
Prior to 6.0 the eventing settings were added to the config as part of
the online upgrade sequence. As the minimum upgrade version has been
bumped to 6.0 there is no longer a need for doing this via online
upgrade. Instead the values are part of the default config.

Change-Id: Ifbed8553fa9698d29005f4a13d73b0cd56e53409
Reviewed-on: http://review.couchbase.org/c/ns_server/+/144749
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>