MB-43269 Remove ssl_services_setup upgrade
Prior to 6.0 the client_cert_auth information was added to the config
as part of the online upgrade sequence. As the minimum upgrade version
has been bumped to 6.0 this can now be a part of the default config.

In addition, get_user_name_from_client_cert should no longer check the
compat mode.

Change-Id: Ic8eb41b9b2f34189e4720a5ed4aff43ace353a54
Reviewed-on: http://review.couchbase.org/c/ns_server/+/144750
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>