implementation of quorum loss chronicle failover
In case of unsafe failover the nodes are failed over in chronicle
first, then the regular failover routine is done and at the end
the nodes will be removed (not deactivated) so they cannot be
added back.

In case of unfinished failover, the key $failover_opaque will
contain the list of nodes that were unsuccessfully failed over.
The next failover will have to include these nodes, so the
failover should be properly finished.

Change-Id: Ib63b80c9556b659ab32793ba0c37a3126b5cfe41
Reviewed-on: http://review.couchbase.org/c/ns_server/+/146394
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>