Ensure that cbcollect_info doesn't crash if it can't walk the data dir
Currently cbcollect_info will crash if it can't walk the data directory
in add_bucket_info(). This is undesirable as it means that any logs that
it might be able to collect will be packaged into a collection. Instead
we should catch the IO exception that is thrown log the fact that this
happened and try and collect as much information about the node as
possible.

Change-Id: I39990efe7af3ae520ed25d475935755f378191f4
Reviewed-on: http://review.couchbase.org/c/ns_server/+/146423
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>