Refactor cb_epmd:port_please code in order to ...
... make it possible to get the port for an arbitrary address family
and encryption setting. Introduce cb_epmd:get_port/4 function.

port_please is used by dist modules for geting the port of a remote
node for currently cofigured settings (afamily and encryption).
But for self testing purposes we want to be able to extract TLS port
of a node even when encryption is disabled, and vise versa,
to extract TCP port when encryption is enabled.

Change-Id: Ia03f7442e9b2ce15ca10b4713ba94ed6db61a483
Reviewed-on: http://review.couchbase.org/c/ns_server/+/146779
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sam Cramer <sam.cramer@couchbase.com>