atomic topology change operations
topology mutations are done in 2 steps:
1. mutate some keys in kv
2. change chronicle topology

If something goes wrong we don't want to end up in state when the
kv mutation is done and chronicle mutation is not

This patch joins 2 steps in one transaction using special key:
unfinished_topology_operation which exists for the duration of
the transaction and keeps the information about the operation
being done

Topology mutation in this commit is performed this way:
1. Grab chronicle lock
2. Check if unfinished_topology_operation exists and execute
   repair operation. Repair operation is executed if the current
   operation doesn't match the one in the key, otherwise we just
   proceed with the current operation.
3. Perform the requested operation both on kv and on chronicle
   topology.
4. remove unfinished_topology_operation key fi the lock matches

Any lock mismatch causes chronicle_master to crash and the call
to chronicle master is retried 3 times if crash happens. The idea
is that eventually there will be a single chronicle_master in
the cluster and therefore no lock conflicts will occure.

Recovery operation for add_replica is remove_peer, because it reflects
the current recovery code in ns_cluster that calls shun() if add_node
fails. For all other operations we just repeat the operation.

If the retries are exceeded we might end up with unfinished transaction.
The janitor for such situations comes in the next commit.

Change-Id: I4038ab21ccad2f971c36448186eb8f233f9de6c5
Reviewed-on: http://review.couchbase.org/c/ns_server/+/147125
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>